Player turns should return character maps, not word maps.
I'm making this decision on the basis of how I plan to
implement the bot - via brute force, more or less.
Rather than forcing the bot to construct the characters
it chooses to add into words, it willb e simpler to
pass back characters. 

The same could be said of the user, actually - rather
than forcing the UI to compose words, it should just
pass back the characters added by the player.