Removed AbstractPlayer in favor of BasePlayer.
I made this decision to simplify the interface for
players, which jammed together things that the game
coordinator should keep track of (like score), which
the player shouldn't be able to freely modify.

This also caused me to remove the tests for 
AbstractPlayer, since everything checked by those tests
is going to be moved to the game anyway.