Wrote a FileConfig for loading from the filesystem, and a game script.
The FileConfig is easier to use during development, since the game script
added in this commit doesn't use a JAR, because the script is run from
the SBT console.

Since the ResourceConfig and FileConfig are similar, I refatored a new
class, StreamConfig, which is generates the Config information using
Java's Scanner.