Extracted TurnBuilder from TerminalCommandProcessor
The purpose of TurnBuilder is to group together functionality for turn
editing - in the TerminalCommandProcessor (and future GUI), the player
interacts with the TurnBuilder's "shadow board" and "shadow rack," which
allows for adding and removing tiles while keeping the rack in sync.

This shrinks TerminalCommandProcessor quite a bit, and will make it easier
to write the GUI, since the turn editing won't have to be copied.