Wrote some wrappers for some of AWT's events.
This is less tedious than subclassing ActionListener or MouseListener,
since I can use a single function literal (and pattern matching in the
case of MouseListener) rather than having to implement a method.

I should have added this before the RackView before, but I don't want to
meddle with history at the moment.