Added multiple UI modes for different stages of the game.
Adding this to the RackView required making the callbacks that the
RackView sets swappable - the RackCallbackSuite trait and its
implementations were added to avoid having to add a bunch of branches
to the existing callbacks.

I also added some code to the RackView test script, for testing the
different modes.