Add special colors for the center tile
Since it is necessary to start branching out from the center tile on the
first turn, it was difficult to start playing using the Swing UI, since
there was no indication which tile was the center. This commit changes
the color of the center tile, so that it is easier to see.