Add a main class for the Swing UI
The new main class, GameSetup, allows the player to put in their name
and choose how many AI players they want, before starting the game. I
also had to edit the build file, so that this class is recognized as the
main class when building JARS via `sbt assemble`.

(There is also a minor change in the build file, which makes the tests
more verbose when the fail. This was added in a while ago, but I never
checked it in previously since it didn't seem to merit a complete
commit.)